Update aquarium-salinity.yaml
FIXED!

Lots of smoothing on the temperature voltage side with EC & Salinity independently calculated from both the internal probe and an external reference temperature that's provided by Home Assistant. 

Ran for over 12 hours with no drift from reference. 

I'll let it run for a week and if it remains as stable as it is, the external ref temp can be removed, though it is useful for both sanity checking as well as to specify the temperature when calibrating to a 53 uS/cm